requirements.txt
fastapi==0.115.0
uvicorn==0.30.6

torch==2.4.1
torchvision==0.19.1
torchxrayvision==1.3.0

numpy==1.26.4
pandas==2.2.3

pydicom==3.0.1
Pillow==10.4.0

python-multipart==0.0.9

supabase==2.9.1

python-dotenv==1.0.1